% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box7.r
\docType{data}
\name{box7}
\alias{box7}
\alias{box8}
\alias{lha.NATL}
\alias{lha.WORLD}
\alias{wgridspp}
\title{Datasets used in SatTagSim of 7-box schema of the north Atlantic}
\format{SpatialPointsDataFrame}
\usage{
data(box7)
data(box8)
data(lha.w)
data(lha.natl)
data(wgridspp)
}
\description{
\itemize{
\item 7 box schema used in the Kerr et al. operational model
\item 8 box schema modified from Kerr et al. operational model
\item Longhurst areas for North Atlantic and World
\item 10 degree boxes
}
}
\examples{
library(fields)
data(box7)
plot(box7, border = 'salmon', lwd=2)
world(add = T)
}
\keyword{datasets}

