% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col2cpt.R
\name{col2cpt}
\alias{col2cpt}
\title{Make a GMT color ramp file (.cpt)}
\usage{
col2cpt(nmin = 0, nmax = 1, clist = c("lightcyan", "royalblue", "blue",
  "lemonchiffon", "orange", "red"), outfile = "mycol.cpt", length = 100)
}
\arguments{
\item{nmin}{numeric minimum of the ramp}

\item{nmax}{numeric maximum of the ramp}

\item{clist}{catenated color list}

\item{outfile}{outpur filename (.cpt)}

\item{length}{number of segments between min and max}
}
\value{
a .cpt file
}
\description{
Make a GMT color ramp file (.cpt)
}
\details{
I love GMT. No other reason for this to be in here.
}
\examples{

col2cpt(nmin = 0, nmax = 40, clist = c("lightcyan", "royalblue", "blue", "lemonchiffon", "orange", "red"),  outfile = 'my_colors.cpt', length=100)
}
\references{
Wessel, P., Smith, W.H.F., Scharroo, R., Luis, J., and Wobbe, F. 2013. Generic Mapping Tools: Improved Version Released. Eos, Transactions American Geophysical Union 94(45): 409–410. doi:10.1002/2013EO450001.
http://gmt.soest.hawaii.edu/doc/latest/index.html
}
\author{
Benjamin Galuardi
}
