% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.allpar.r
\name{get.allpar}
\alias{get.allpar}
\title{Get u and v parameterss}
\usage{
get.allpar(tracks = nsfish, parallel = FALSE)
}
\arguments{
\item{tracks}{A group of tracks returned from Kalman filter estimation. Columns must be in the following order: Day, Month, Year, Longitude, Latitude}

\item{parallel}{Logical. if parallel, uses plyr built in parallel capabilities}
}
\value{
a data frame of TagID, Month, u, v and a default value of D (diffusion) for each track for each month. This result is a pre requisite for \code{\link{merge.par}}
}
\description{
Function to get monthly u and v advection parameters for a group of tracks
}
\seealso{
\code{\link{merge.par}} \code{\link{get.uv}}
}
\author{
Benjamin Galuardi
}
