% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.box.vals.R
\name{get.box.vals}
\alias{get.box.vals}
\title{Get overlay of boxes from a track}
\usage{
get.box.vals(dat, boxes)
}
\arguments{
\item{dat}{a dataframe track with lon, lat and Month columns}

\item{boxes}{spatial polygons of overlay strata}
}
\description{
Get overlay of boxes from a track
}
\seealso{
\code{\link{get.first.box}}
}
