% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.kfD.r
\name{get.kfD}
\alias{get.kfD}
\title{Get D from KF estimates}
\usage{
get.kfD(tracks = nsfish)
}
\arguments{
\item{tracks}{Track(s) returned from Kalman filter estimation. Columns must be in the following order: Day, Month, Year, Longitude, Latitude and include D, sd(D), Month and TagID (exactly as spelled and capitalized)}
}
\value{
a data frame of Month, TagID,  D, Dsd, nrec (number of records)
}
\description{
Function to add diffusion (D) and sd(D) from Kalman filter estimations. D and sd(D) must be added  to the track as columns beforehand. D is taken from kalmn filter estimte or is a fixed value used in the state space estimation process
}
\examples{
see vignette
}
\seealso{
\code{\link{merge.par}} \code{\link{get.uv}} \code{\link{get.allpar}}
}
\author{
Benjamin Galuardi
}
