% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.start.pts.r
\name{get.start.pts}
\alias{get.start.pts}
\title{Get starting points
Get starting points for simulations based on distribution of observed locations for a given month.}
\usage{
get.start.pts(dat, n = 100, months = 1:12, mask = rmask,
  posnames = c("coords.x1", "coords.x2"))
}
\arguments{
\item{dat}{input data. see \link{\code{nsdata}}}

\item{n}{number of starting points}

\item{posnames}{names of the longitude and latitude columns in the data frame}
}
\value{
a list of starting points
}
\description{
Get starting points
Get starting points for simulations based on distribution of observed locations for a given month.
}
\details{
The number of starting points is the number of tracks desired for simulation in a given month
}
\examples{
library(fields)
data(nsfish)
spts = get.start.pts(nsfish)
world(xlim = c(-100, 0), ylim = c(20, 50))
lapply(spts, points, pch = 19, cex = .4)
}
