% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.trans.prob.r
\name{get.trans.prob}
\alias{get.trans.prob}
\title{Transition probability matrix generator}
\usage{
get.trans.prob(datbox, nyears = 100, adims = c(7, 7, 4), perc = TRUE, ...)
}
\arguments{
\item{datbox}{data frame returned from \code{\link{get.first.box}}}

\item{nyears}{dummy variable that sets up a dataframe to be filled. Should be greater than the number of years simulated}

\item{adims}{array dimensions. For a 7-box area and 4 seasons, c(7, 7, 4)}

\item{perc}{logical. Return results as a percentage of the row total}
}
\value{
a list of transition probabilities for each time period. The row index represents the previous area and the column represnts the current area for the given time period.
}
\description{
This function is designed to run on a data frame of many simulated tracks. It loops through each one to get transitions and sums the probabilities for the entire set of simulated tracks. This function uses the first area occupied at the beginning of the time period, and transitions to the first occupied area in the next time period.
}
\examples{
see vignette
}
