% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.uv.r
\name{get.uv}
\alias{get.uv}
\title{Get u and v subfunction}
\usage{
get.uv(track = nsfish[1:10, ])
}
\arguments{
\item{track}{track returned from Kalman filter estimation. Columns must be in the following order: Day, Month, Year, Longitude, Latitude}
}
\value{
u and v advection for the track, or section of track.
}
\description{
Get advective parameters from an estimated track or section of track
}
\examples{
none. Typically used within other functions
}
\author{
Benjamin Galuardi
}
\seealso{
\code{\link{merge.par}} \code{\link{get.kfD}} \code{\link{get.allpar}}
}

