% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_first_box_mod.r
\name{get_first_box_mod}
\alias{get_first_box_mod}
\title{Get First Box}
\usage{
get_first_box_mod(simdat, syear = 2000, boxes = box7, seas.len = 90)
}
\arguments{
\item{simdat}{list of simulated tracks from \code{\link{make.sim.track.par}}}

\item{syear}{start year; a dummy variable for date determination}

\item{boxes}{spatial polygons strata for transition determination}

\item{seas.len}{length of season. typically 90 days but can be adjusted for monthly or other time intervals}
}
\description{
Get First Box
}
\examples{
Modernized First Box Function (Alias)

}
