% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.rasbox.r
\name{make.rasbox}
\alias{make.rasbox}
\title{Rasterize a spatial structure
returns either a raster or cell matrix based on a spatial box structure}
\usage{
make.rasbox(boxes = box11, rrow = 26 * 5, rcol = 29 * 5, raster = T)
}
\arguments{
\item{boxes}{spatial polygons}

\item{rrow}{number of rows for output}

\item{rcol}{number of columns for output}

\item{raster}{Logical. If False, returns a matrix}
}
\value{
a raster or matrix
}
\description{
Rasterize a spatial structure
returns either a raster or cell matrix based on a spatial box structure
}
\examples{
par(mfrow = c(3,1))
data(box11)
r = make.rasbox(box11)
m = make.rasbox(box11, raster = F)
plot(box11, border = 2); world(add = T)
plot(r); world(add = T)
world(add=F); image.plot(m$lon, m$lat, m$box, add=T)
}
