% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.seas.idx.r
\name{make.seas.idx}
\alias{make.seas.idx}
\title{Make seasonal index of tracks}
\usage{
make.seas.idx(btracks)
}
\value{
a list of logical indices by season
}
\description{
Make seasonal index of tracks
}
\details{
btracks must have a 'Month' column
}
\examples{
see vignette
}
