% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.sim.raster.r
\name{make.sim.raster}
\alias{make.sim.raster}
\title{Make sim raster}
\usage{
make.sim.raster(simdat, xmn = -100, xmx = 30, ymn = 0, ymx = 60,
  boxsize = 60)
}
\arguments{
\item{simdat}{Simulated track returned from \code{\link{make.sim.track.par}}}

\item{xmn}{longitude minimum}

\item{xmx}{longitude maximum}

\item{ymn}{latitude minimum}

\item{ymx}{latitude maximum}

\item{boxsize}{size of pixel in nautical miles (minutes)}
}
\value{
raster; cell values are counts of positions per pixel
}
\description{
Make a raster of simulation results
}
\examples{
see Vignettes

}
\author{
Benjamin Galuardi
}
