% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.sim.track.par2.r
\name{make.sim.track.par2}
\alias{make.sim.track.par2}
\title{Simulate tracks in parallel (Legacy wrapper)}
\usage{
make.sim.track.par2(
  par_array = par_array,
  simorder = simorder,
  sp = spts,
  bath = bath,
  sstmat = sstmat,
  boxmat = boxmat,
  seaslen = 30,
  sstol = 2,
  mcoptions = setup.parallel(),
  ...
)
}
\description{
Simulate tracks in parallel (Legacy wrapper)
}
\examples{
see vignette
}
\author{
Benjamin Galuardi
}
