% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdata.r
\docType{data}
\name{nsfish}
\alias{nsfish}
\title{Nova Scotia Atlantic bluefin tuna}
\format{A SpatialPointsDataFrame with 4482 rows and 26 variables: 31 individual tagged fish.}
\source{
\url{www.tunalab.org}

\url{http://www.nrcresearchpress.com/doi/abs/10.1139/f10-033}
}
\usage{
nsfish
}
\description{
Dataset of PSAT tagged Atlantic bleufin tuna from Galuardi et al. 2010
}
\details{
Kalman filter estimated tracks of 31 Atlantic bleufin tuna including Diffusive terms used in the track estimation
}
\keyword{datasets}
