% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.boxtrans.r
\name{plot.boxtrans}
\alias{plot.boxtrans}
\title{plot.boxtrans}
\usage{
\method{plot}{boxtrans}(
  boxtrans,
  palette = "BrBG",
  text.col = "white",
  text.size = 6,
  strip.text = 15,
  axis.text = 20
)
}
\arguments{
\item{boxtrans}{transition matrices returned from \code{\link{get.trans.prob}}}

\item{palette}{continuous color palette see \code{\link{RColorBrewer}}}

\item{text.col}{text color}

\item{text.size}{text size of cells}

\item{strip.text}{text size of panel strips}

\item{axis.text}{text size of axes}
}
\value{
a four panel, seasonal plot (ggplot)
}
\description{
plot.boxtrans
}
