% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.seas.change.r
\name{plot.seas.change}
\alias{plot.seas.change}
\title{Plot percent change from previous season}
\usage{
plot.seas.change(sr, box = NULL)
}
\arguments{
\item{sr}{list of seasonal raster results from}

\item{box}{overlay of boxes for transfer areas}
}
\value{
a plot
}
\description{
Plot percent change from previous season
}
\examples{
# after running make.sim.track.par
simdatdf = ldply(simdat)
simdatdf$seas[simdatdf$Month\%in\%c(1,2,3)] = 1
simdatdf$seas[simdatdf$Month\%in\%c(4,5,6)] = 2
simdatdf$seas[simdatdf$Month\%in\%c(7,8,9)] = 3
simdatdf$seas[simdatdf$Month\%in\%c(10,11,12)] = 4
sr = dlply(simdatdf, 'seas', function(x) make.sim.raster(x, boxsize=60))
plot.seas.change(sr)
}

