% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.uv.density.r
\name{plot.uv.density}
\alias{plot.uv.density}
\title{Diagnostic ggplots of advection}
\usage{
\method{plot}{uv.density}(
  allpar,
  save = FALSE,
  fname = NULL,
  ulims = c(-50, 50),
  vlims = c(-50, 50),
  ...
)
}
\arguments{
\item{allpar}{data frame returned from \code{\link{get.allpar}}}

\item{save}{should the plot be saved}

\item{fname}{filename}

\item{ulims}{plot limits in u (x) direction}

\item{vlims}{plot limits in v (y) direction}

\item{...}{additional arguements to ggsave}
}
\value{
a ggplot
}
\description{
Diagnostic ggplots of advection
}
\examples{
see vignette
}
