% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmask2array.r
\name{rmask2array}
\alias{rmask2array}
\title{Function to tranfrom a raster brick of sst to a 3d array and orient properly}
\usage{
rmask2array(rmask)
}
\arguments{
\item{rmask}{raster mask of monthly suitable temperatures fro Atlantic bluefin tuna}
}
\value{
a 3D array
}
\description{
Function to tranfrom a raster brick of sst to a 3d array and orient properly
}
\examples{
data(rmask)
sst = rmask2array(rmask)
}
