% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.parallel.r
\name{setup.parallel}
\alias{setup.parallel}
\title{Setup parallel processing
Setup parallel processing for simulations}
\usage{
setup.parallel()
}
\value{
multicore options for prallel processing
}
\description{
Setup parallel processing
Setup parallel processing for simulations
}
\examples{
mcoptions = setup.parallel()

}
