% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_tracks_par.R
\name{sim_tracks_par}
\alias{sim_tracks_par}
\title{Simulate tracks in parallel}
\usage{
sim_tracks_par(
  par_array = par_array,
  simorder = simorder,
  sp = spts,
  bath = bath,
  sstmat = sstmat,
  boxmat = boxmat,
  seaslen = 30,
  sstol = 2,
  mcoptions = setup.parallel(),
  ...
)
}
\description{
Simulate tracks in parallel
}
