% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simm.kf.R
\name{simm.kf}
\alias{simm.kf}
\title{Simulate KF}
\usage{
simm.kf(n = 100, u = c(0, 1), v = c(0, 1), D = c(100, 50), sp = c(-70,
  40), ulim = c(-50, 50), vlim = c(-50, 50), Dlim = c(0, 5000))
}
\arguments{
\item{n}{number of points to simulate}

\item{u}{vector of mean and standard deviation of advection in u (longitude) direction, in nautical miles/day}

\item{v}{vector of mean and standard deviation of advection in v (latitude) direction, in nautical miles/day}

\item{D}{vector of mean and standard deviation of Diffusion, in nautical miles^2/day}

\item{sp}{vector of starting point (lon, lat)}
}
\value{
a two column matrix of longitude and latitude
}
\description{
Correlated random walk simluator based on advection diffusion process equation
}
\details{
uses truncated normal distributions. Bounds are:
\tabular{lll}{
Parameter \tab lower bound \tab upper bound \cr
u \tab -50  \tab 50  \cr
v \tab -50  \tab 50  \cr
D \tab 0  \tab 5000
}
}
\examples{
# Three examples of differing advection and Diffusion parameter combinations
par(mfrow=c(1,3))
plot(simm.kf(n = 1000, u = c(5,1), v = c(5,1), D = c(10,1), sp = c(-70,40)), typ = 'o', pch = 19, col =2, xlab = 'lon', ylab = 'lat')
plot(simm.kf(n = 1000, u = c(10,1), v = c(0,1), D = c(10,1), sp = c(-70,40)), typ = 'o', pch = 19, col =3, xlab = 'lon', ylab = 'lat')
plot(simm.kf(n = 1000, u = c(0,0), v = c(0,0), D = c(5000,1000), sp = c(-70,40)), typ = 'o', pch = 19, col =4, xlab = 'lon', ylab = 'lat')

}
\references{
Sibert, J. R., Musyl, M. K. and Brill, R. W. 2003. Horizontal movements of bigeye tuna (Thunnus obesus) near Hawaii determined by Kalman filter analysis of archival tagging data. Fish. Oceanogr. 12(3): 141?151.
Calenge, C. (2006) The package adehabitat for the R software: a tool for the analysis of space and habitat use by animals. Ecological Modelling, 197, 516-519
}
\seealso{
\code{\link{make.sim.track.par}}
}
\author{
Benjamin Galuardi
}
