% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.first.box.R
\name{get.first.box}
\alias{get.first.box}
\title{Get first occupied area in a time frame}
\usage{
get.first.box(simdat, syear = 2000, boxes = box7, seas.len = 90, nyears = 2)
}
\arguments{
\item{simdat}{list of simulated tracks from \code{\link{make.sim.track.par}}}

\item{syear}{start year; a dummy variable for date determination}

\item{boxes}{spatial polygons strata for transition determination}

\item{seas.len}{length of season. typically 90 days but can be adjusted for monthly or other time intervals}
}
\value{
data frame with Year season previous box (strata) matrix transition and current box (strata). nuumbers of previous box  and current box refer to the polygon order in the the boxes parameter
}
\description{
Get first occupied area in a time frame
}
\examples{
see vignette
}
\seealso{
\code{\link{get.trans.prob}} \code{\link{box7}}
}
