% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_first_box_mod.r
\name{get_first_box_mod}
\alias{get_first_box_mod}
\title{Get First Box}
\usage{
get_first_box_mod(simdat, syear = 2000, boxes = box7, seas.len = 90)
}
\arguments{
\item{seas.len}{}
}
\description{
Get First Box
}
