% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.par.array.r
\name{make.par.array}
\alias{make.par.array}
\title{Make spatio-temporal array of movement parameters
This function generates monthly advection and diffusion mean and standard deviation for each area in the spatial strata used. Missing months default to the previous months value. Missing areas should be filled in by the user.}
\usage{
make.par.array(
  tracks,
  inbox,
  rasbox = NULL,
  rrows = 26 * 5,
  rcols = 29 * 5,
  use_wts = NULL,
  missvec = NULL,
  fillvec = NULL
)
}
\arguments{
\item{tracks}{spatial points data frame of tracks}

\item{inbox}{spatial polygons (shapefile) of areas}

\item{rasbox}{optional raster version of spatial polygons for areas. If included, rrows and rcols may be NULL}

\item{rrows}{number of rows for raster creation}

\item{rcols}{number of columns for raster creation}

\item{use_wts}{T/F for using weights}

\item{missvec}{vector of missing rows (areas) in final output. This can be obtained by running once with this function left as NULL}

\item{fillvec}{vector of replacement rows for missvec. This is a subjectie decision by the user and is dependent on the spatial strata used}
}
\value{

}
\description{
Make spatio-temporal array of movement parameters
This function generates monthly advection and diffusion mean and standard deviation for each area in the spatial strata used. Missing months default to the previous months value. Missing areas should be filled in by the user.
}
\details{
The default size raster created is 130 degrees latitude and 145 degrees longitude, with 5 degree cells. This should be specific to the spatial strata used (e.g. the 11 box model) and should ideally have cells that split the areas along ploygon lines.
}
