% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.sim.track.par2.r
\name{make.sim.track.par2}
\alias{make.sim.track.par2}
\title{Simulate tracks in parallel}
\usage{
make.sim.track.par2(par_array = par_array, simorder = simorder, sp = spts,
  bath = bath, sstmat = sstmat, boxmat = boxmat, seaslen = 30,
  sstol = 2, mcoptions = setup.parallel(), ...)
}
\arguments{
\item{par_array}{array of temporal and spatially explcit simulation parameters @seealso xxxx}

\item{sp}{list of starting points}

\item{sstmat}{list of lon, lat and 3D matrix of monthly sea surface temperature. see details}

\item{boxmat}{raster of spatial strata used when generating movement parameters @seealso xxxx}

\item{seaslen}{length of each month to simulate. Defaults to 30 (days per month)}

\item{sstol}{tolerance for sst matching. see details}

\item{mcoptions}{This is a hidden variable that must be present to run in parallel}

\item{morder}{month order for simulation (e.g. start in June vs start in January)}
}
\value{
List of simulated tracks. Tracks are data frames with columns: lon lat month
}
\description{
Function to simulate a single track using monthly parameters, bathymetry and World Ocean Atlas SST in parallel
}
\details{
sstmat in this package is a running average of index suitable temperature for each month. Each month has values 0-3. A value of 3, for example, represents areas suitable in the current month, previous month and next month. The sstol flag can make matching criteria more or less restrictive. The default value means that at least two months of suitability, centered on the current month, must be valid for a simulation point.
Tracks may be generated by increasing morder (e.g. rep(morder, 2) for a 24 month track)
Function uses mcoptions from \code{\link{setup.parallel}} to run in parallel.
}
\examples{
see vignette
}
\author{
Benjamin Galuardi
}
