% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.par.r
\name{merge.par}
\alias{merge.par}
\title{Merge u,v,D parameters
Function to merge UV and D parameter data frames together}
\usage{
\method{merge}{par}(uvpar, Dpar, track = NULL, return.mean = F)
}
\arguments{
\item{uvpar}{dataframe of u and v advection parameters from \code{\link{get.allpar}}}

\item{Dpar}{dataframe of u and v advection parameters from \code{\link{get.allpar}}}

\item{track}{tagID index. If merging a single (or group) track/tagid, this is ithe tagid to merge. Otherwise, all records are merged}

\item{return.mean}{logical; should a mean set of values per month be returned? If False, mean values for each month, for each fish, are returned.}
}
\value{
data frame of: Month, u, v, D, sd.u, sd.v, sd.D
}
\description{
Merge u,v,D parameters
Function to merge UV and D parameter data frames together
}
\examples{
data("nsfish")
uvpar = get.allpar(as.data.frame(nsfish))
Dpar = get.kfD(as.data.frame(nsfish))
simpar = merge.par(uvpar, Dpar, return.mean=T)

}
\author{
Benjamin Galuardi
}
\seealso{
\code{\link{get.kfD}} \code{\link{get.uv}} \code{\link{get.allpar}}
}

