% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modern_process_tracks.r
\name{modern_process_tracks}
\alias{modern_process_tracks}
\title{Get parameters modern method
alternate way of getting movement parameters 
Experimental only..}
\usage{
modern_process_tracks(
  tracks,
  inbox,
  rasbox = NULL,
  rrows = 26 * 5,
  rcols = 29 * 5,
  use_wts = NULL,
  missvec = NULL,
  fillvec = NULL
)
}
\arguments{
\item{fillvec}{}
}
\description{
Get parameters modern method
alternate way of getting movement parameters 
Experimental only..
}
\examples{
simpar_dl_test  =  modern_process_tracks(tracks  =  dlfish_sp
, inbox  =  my_boxes
, rasbox  =  NULL
, rrows  =  26*5
, rcols  =  29*5
, missvec  =  c(1)
, fillvec  =  c(2)
, use_wts  =  NULL
)
}
